requirements.txt
Django==4.1
django-environ==0.11.2
django-filter==23.2
djangorestframework==3.14.0
django-cors-headers==4.2.0
mysqlclient==2.1.1
wrf-python==1.3.4.1
netCDF4==1.6.4
matplotlib==3.5.0
geojsoncontour==0.4.0
geojson==3.0.1
numpy==1.25.2
